Fixes #131 - Exchange no longer uses 100% CPU
The data exchange had an infinite loop for adapters to receive. If an adapter did not
have a receive method then this was never blocking and used 100% CPU. To fix this, a
check of the return value from receive was added. If the return value is None then it
is assumed that the adapter does not need a receive method and the adapter thread exits